refactor: extract shared wait_for_ssh into stubs/common/ssh_utils.py
Deduplicate the wait_for_ssh helper that was copy-pasted across 6 AWS
stubs (bare_metal and vm start/reboot/power_cycle/reinstall). VM stubs
use shorter poll intervals (30 attempts × 10s) vs bare-metal defaults
(60 × 15s).

Made-with: Cursor
